Add an ant task for pre-compiling JSPs (#68)
* Add task that will compile JSPs in a war

compiling JSPs at runtime can be expensive so compiling them at build
time is preferable. Liberty doesn’t do this though so simplest solution
is to simply have an ANT task that starts the server, waits for the
compilation to succeed and then adds them into the war file.

* Enhance task to compile jsps standalone

Enhance the compile JSP task so in addition to compiling jsps in a war
file it can compile jsps in a source dir.

* Update the documentation for the new task

Provide documentation for the new task.

* Fix broken link in README.md

Fix the broken link in the readme to the compileJSPs task.

* Update tests so they run without assuming setup for my machine

Update the tests to install Liberty so it’ll run standalone. Also write
files to target dir so it wont get accidentally checked in.

* Update to put the jsps in the root of WEB-INF/classes

The JSP Engine in Liberty does not load the jsp class files if they are
in the com/ibm/_jsp dir structure despite being in that package. Weird.

* Minor updates

A few minor, but important updates:

1) Add an After method on the test so the classes dir gets cleaned up
after the test runs
2) Update the temp folder used to be named after the task
3) Move the checkFeaturesExist call into the try/finally so if the
checkFeaturesExist step fails the server is still shut down.

* Remove unused  package imports

Unused package imports

* update Liberty version string for download to one that works